[step2] bug fix: 나눗셈 소수점 누락되는 버그 수정, pr 반영
- pr 리뷰 반영 : https://github.com/next-step/kotlin-racingcar/pull/431#discussion_r864564175

+++ src/main/kotlin/step2/calculator/Calculator.kt
@@ -4,14 +4,14 @@ import step2.calculator.utils.CustomStringParser
 
 class Calculator {
 
-    fun calculate(input: String?): Long {
+    fun calculate(input: String?): Double {
         val strings = CustomStringParser.parse(input)
         val token = CustomStringParser.removeEmptyStrings(strings)
 
         return _calculate(token)
     }
 
-    private fun _calculate(token: List<String>): Long {
+    private fun _calculate(token: List<String>): Double {
         val operator = getOperator(token)
         val (first, second) = Pair(token[FIRST_NUMBER_POSITION_INDEX], token[SECOND_NUMBER_POSITION_INDEX])
